fix(plugin-sdk): expose delivery hints without utility imports
