+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-2ac33288de5918da052ee7f1715294135aa99fc829666cc9490d1cd3e17f8a10  plugin-sdk-api-baseline.json
-a37ed8fc317d439f33425058c2be7114d8278f7e9ed82fb869ef783ae73bca26  plugin-sdk-api-baseline.jsonl
+40b3c841849fbc29938a3bbb990e28a5db30142941c8ef0c081a94cee4c78331  plugin-sdk-api-baseline.json
+40ee8e1bbf112e768d4944776443f90b2441b02e3e950726e4112015cd106108  plugin-sdk-api-baseline.jsonl
