+++ scripts/lib/plugin-sdk-doc-metadata.ts
@@ -102,6 +102,9 @@ export const pluginSdkDocMetadata = {
   "provider-oauth-runtime": {
     category: "provider",
   },
+  "message-tool-delivery-hints": {
+    category: "runtime",
+  },
   "provider-selection-runtime": {
     category: "provider",
   },
