+++ src/auto-reply/reply/delivery-hints.ts
@@ -2,4 +2,4 @@ export {
   LEGACY_MESSAGE_TOOL_DELIVERY_HINTS,
   MESSAGE_TOOL_DELIVERY_HINTS,
   MESSAGE_TOOL_ONLY_DELIVERY_HINT,
-} from "../../plugin-sdk/text-utility-runtime.js";
+} from "../../plugin-sdk/message-tool-delivery-hints.js";
